fix: address Claude review on PR #65 — boolean override, line_spacing type, shared-style mutation, test gaps (JON-91)
Resolves the five issues flagged in the Claude bot review on PR #65:

1. extract.py: the boolean falsy-check on keep_together / keep_with_next /
   page_break_before silently dropped explicit False overrides. Now use
   `is not None` so False survives round-trip (e.g. overriding a Heading 1
   default of True to False).

2. models.py + extract.py: line_spacing is Optional[float], not Optional[int].
   python-docx returns 1.5 for proportional spacing — int() was truncating to 1.
   The cast is removed and the type widened.

3. reconstruct.py: font_name / font_size in block_style are now applied
   per-run, not on para.style.font.*. The style object is shared across all
   paragraphs with the same docx_style, so the previous code cross-contaminated
   fonts between paragraphs whenever JON-91 caused docx_style to be applied.

4. tests: test_missing_custom_style_falls_back_to_normal now asserts that the
   "style not found, falling back to Normal" UserWarning actually fires. The
   assertion calls _apply_block_formatting directly since CliRunner.invoke
   does not propagate warnings out of its subprocess scope.

5. tests: every CliRunner.invoke(main, ["extract", ...]) now assigns
   extract_result and asserts exit_code == 0. Silent extraction failures
   previously surfaced only as confusing build errors downstream.

Also addresses:
- models.py: corrected the misleading "EMUs" comment to reflect that
  line_spacing may be a float and the three flags are bool.
- reconstruct.py: documented why "Normal", "Table", and "TextBox" are
  skipped in the docx_style application gate.
- tests: added test_old_styles_json_without_new_fields_builds to verify
  pre-JON-91 archives still build with the new code (forward compatibility).

+++ src/sidedoc/models.py
@@ -113,13 +113,15 @@ class Style:
     italic: Optional[bool] = None
     underline: Optional[bool] = None
     table_formatting: Optional[dict[str, Any]] = None  # For tables: column_widths, table_alignment, table_style, cell_styles
-    # Paragraph format properties (stored in EMUs / English Metric Units)
+    # Paragraph format properties. Indents/spacing in EMUs (int); line_spacing
+    # is a float for proportional spacing (1.5 = 1.5x lines) or an int EMU for
+    # exact spacing; the flags are bool.
     left_indent: Optional[int] = None
     right_indent: Optional[int] = None
     first_line_indent: Optional[int] = None
     space_before: Optional[int] = None
     space_after: Optional[int] = None
-    line_spacing: Optional[int] = None
+    line_spacing: Optional[float] = None
     keep_together: Optional[bool] = None
     keep_with_next: Optional[bool] = None
     page_break_before: Optional[bool] = None
